+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The *alited* provides the rich and comprehensive set of features, some are uniqu
 # LAST CHANGES
 
 
-Version `v1.8.13 (3 Jun'26)`
+Version `v1.8.13 (1 Jul'26)`
 
   - BUGFIX: no theming of book-mark's context menu (clam theme)
   - BUGFIX: Saved Favorites: list of files is ruined at "Change" action
@@ -17,6 +17,7 @@ Version `v1.8.13 (3 Jun'26)`
   - NEW   : Preference/General/Test: added to preview: left tool bar & menu bar
   - CHANGE: Projects: if current project is chosen, no actions (just close Projects)
   - CHANGE: some clearance
+  - CHANGE: packages: apave/e_menu 4.9.0
 
 
 Version `v1.8.12 (25 Mar'26)`